+++ opal/caregivers/factories.py
@@ -1,0 +1,18 @@
+"""Module providing model factories for caregiver app models."""
+
+from factory import Sequence, SubFactory
+from factory.django import DjangoModelFactory
+
+from opal.users.factories import Caregiver
+
+from . import models
+
+
+class CaregiverProfile(DjangoModelFactory):
+    """Model factory to create [opal.caregivers.models.CaregiverProfile][] models."""
+
+    class Meta:
+        model = models.CaregiverProfile
+
+    user = SubFactory(Caregiver)
+    legacy_id = Sequence(lambda number: number + 1)